+++ demo/tests/test_integration.py
@@ -391,10 +391,14 @@ class MyForm(FormidableForm):
 class TestValidationEndPoint(FormidableAPITestCase):
 
     url = 'formidable:form_validation'
+    method = 'get'
 
     def setUp(self):
         super(TestValidationEndPoint, self).setUp()
         self.formidable = MyForm.to_formidable(label='title')
+
+    def get_method(self):
+        return getattr(self.client, self.method)
 
     def test_validate_data_ok(self):
         parameters = {
@@ -404,7 +408,8 @@ class TestValidationEndPoint(FormidableAPITestCase):
         session = self.client.session
         session['role'] = 'padawan'
         session.save()
-        res = self.client.get(
+        func = self.get_method()
+        res = func(
             reverse(self.url, args=[self.formidable.pk]),
             parameters, format='json'
         )
@@ -419,7 +424,8 @@ class TestValidationEndPoint(FormidableAPITestCase):
         session = self.client.session
         session['role'] = 'padawan'
         session.save()
-        res = self.client.get(
+        func = self.get_method()
+        res = func(
             reverse(self.url, args=[9999]),
             parameters, format='json'
         )
@@ -432,7 +438,8 @@ class TestValidationEndPoint(FormidableAPITestCase):
         session = self.client.session
         session['role'] = 'padawan'
         session.save()
-        res = self.client.get(
+        func = self.get_method()
+        res = func(
             reverse(self.url, args=[self.formidable.pk]),
             parameters, format='json'
         )
@@ -452,7 +459,8 @@ class TestValidationEndPoint(FormidableAPITestCase):
         session = self.client.session
         session['role'] = 'padawan'
         session.save()
-        res = self.client.get(
+        func = self.get_method()
+        res = func(
             reverse(self.url, args=[formidable.pk]),
             parameters, format='json'
         )
@@ -493,7 +501,8 @@ class TestValidationEndPoint(FormidableAPITestCase):
 
         # The checkbox is checked.
         parameters = {'checkbox': True}
-        res = self.client.get(
+        func = self.get_method()
+        res = func(
             reverse(self.url, args=[formidable.pk]),
             parameters, format='json'
         )
@@ -503,7 +512,8 @@ class TestValidationEndPoint(FormidableAPITestCase):
 
         # The checkbox is NOT checked.
         parameters = {'checkbox': False}
-        res = self.client.get(
+        func = self.get_method()
+        res = func(
             reverse(self.url, args=[formidable.pk]),
             parameters, format='json'
         )
@@ -514,3 +524,8 @@ class TestValidationEndPoint(FormidableAPITestCase):
 class TestValidationFromSchemaEndPoint(TestValidationEndPoint):
 
     url = 'form_validation_schema'
+
+
+class TestValidationWithPostMethod(TestValidationEndPoint):
+
+    method = 'post'