+++ docs/source/_static/specs/formidable.js
@@ -482,7 +482,29 @@ var spec = {
                         }
                     }
                 },
-                "summary": "Validate a form"
+                "summary": "Validate a form (GET method). GET and POST are equivalent, but GET is deprecated."
+            },
+            "post": {
+                "parameters": [
+                    {
+                        "in": "path",
+                        "name": "id",
+                        "required": true,
+                        "type": "integer"
+                    }
+                ],
+                "responses": {
+                    "204": {
+                        "description": "Validation OK"
+                    },
+                    "400": {
+                        "description": "Validation KO",
+                        "schema": {
+                            "$ref": "#/definitions/InputError"
+                        }
+                    }
+                },
+                "summary": "Validate a form (POST method). GET and POST are equivalent."
             }
         }
     },